Contents of demo_UAC2_v073.

+++ src/taskMoboCtrl.h
@@ -13,7 +13,7 @@
 // SWR value x 100, in unsigned int format
 extern uint16_t	measured_SWR;
 
-extern uint16_t measured_Power(uint16_t voltage);
+extern uint32_t measured_Power(uint16_t voltage);
 extern void 	vStartTaskMoboCtrl(void);
 
 #endif
